+++ src/AdventOfCode.Solutions2020/Day17/Coordinate4D.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 namespace AdventOfCode.Solutions2020.Day17;
 
-public record Coordinate4D (int X, int Y, int Z, int W)
+public record Coordinate4D (long X, long Y, long Z, long W)
 {
     public static Coordinate4D operator +(Coordinate4D left, Coordinate4D right)
     {
@@ -11,13 +11,13 @@ public record Coordinate4D (int X, int Y, int Z, int W)
         
     public IEnumerable<Coordinate4D> GetAdjacent()
     {
-        for (int x = -1; x <= 1; x++)
+        for (long x = -1; x <= 1; x++)
         {
-            for (int y = -1; y <= 1; y++)
+            for (long y = -1; y <= 1; y++)
             {
-                for (int z = -1; z <= 1; z++)
+                for (long z = -1; z <= 1; z++)
                 {
-                    for (int w = -1; w <= 1; w++)
+                    for (long w = -1; w <= 1; w++)
                     {
                         if (!(x == 0 && y == 0 && z == 0 && w == 0))
                             yield return new Coordinate4D(x, y, z, w) + this;
